Added `Number` struct to toggle float and double

+++ src/Emuratch.Core.Test/ProjectTest.cs
@@ -12,11 +12,11 @@ public class ProjectTest
 
 		Assert.Equal("Here's sprite name", project.sprites[0].name);
 		Assert.Equal("d1163a80056fa6ba2893e707e61ddc7e", project.sprites[0].costumes[0].assetId);
-		Assert.Equal(174, project.sprites[0].costumes[0].Width);
-		Assert.Equal(74, project.sprites[0].costumes[0].Height);
-		Assert.Equal(0, project.sprites[0].x);
-		Assert.Equal(0, project.sprites[0].y);
-		Assert.Equal(90, project.sprites[0].direction);
+		Assert.Equal(174, (int)project.sprites[0].costumes[0].Width);
+		Assert.Equal(74, (int)project.sprites[0].costumes[0].Height);
+		Assert.Equal(0, (int)project.sprites[0].x);
+		Assert.Equal(0, (int)project.sprites[0].y);
+		Assert.Equal(90, (int)project.sprites[0].direction);
 		Assert.Equal(3, project.sprites[0].blocks.Count);
 		Assert.False(project.sprites[0].isClone);
 		Assert.True(project.stage.isStage);
